Add user endpoints + refactoring (#7)
* Add user entity, repo, service and controller + refactor

* Add endpoint to add policies to a created user

* Modify policy entity to use lombok
