+++ src/main/java/ro/unibuc/hello/controllers/contracts/PolicyCreateRequest.java
@@ -1,0 +1,17 @@
+package ro.unibuc.hello.controllers.contracts;
+
+
+import lombok.*;
+import ro.unibuc.hello.entities.Statement;
+
+import java.util.List;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class PolicyCreateRequest  {
+    private String name;
+    private List<Statement> statements;
+}
