+++ src/main/java/ro/unibuc/hello/controllers/contracts/UserCreateRequest.java
@@ -1,0 +1,15 @@
+package ro.unibuc.hello.controllers.contracts;
+
+import lombok.*;
+
+import java.util.List;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class UserCreateRequest {
+    private String id;
+    private List<String> policies;
+}
