--- src/main/java/ro/unibuc/hello/data/ActionEntity.java
@@ -1,24 +1,0 @@
-package ro.unibuc.hello.data;
-
-import org.springframework.data.annotation.Id;
-
-public class ActionEntity {
-
-    @Id
-    public String actionCode;
-    public String actionDescription;
-
-    public ActionEntity() {}
-
-    public ActionEntity(String code, String description){
-        this.actionCode = code;
-        this.actionDescription = description;
-    }
-
-    @Override
-    public String toString() {
-        return String.format(
-            "Action[cod=%s, description='%s']",
-            actionCode, actionDescription);
-    }
-}
