--- src/main/java/ro/unibuc/hello/data/InformationRepository.java
@@ -1,18 +1,0 @@
-package ro.unibuc.hello.data;
-
-import java.util.List;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * No need to implement this interface.
- * Spring Data MongoDB automatically creates a class it implementing the interface when you run the application.
- */
-@Repository
-public interface InformationRepository extends MongoRepository<InformationEntity, String> {
-
-    InformationEntity findByTitle(String title);
-    List<InformationEntity> findByDescription(String description);
-
-}
