--- src/main/java/ro/unibuc/hello/dto/Action.java
@@ -1,28 +1,0 @@
-package ro.unibuc.hello.dto;
-
-public class Action {
-    private String code;
-    private String description;
-
-    public Action(String code, String description) {
-        this.code = code;
-        this.description = description;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-}
