+++ src/main/java/ro/unibuc/hello/exceptions/EntityAlreadyExistsException.java
@@ -1,0 +1,13 @@
+package ro.unibuc.hello.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value= HttpStatus.CONFLICT, reason="Entity already exists")
+public class EntityAlreadyExistsException extends RuntimeException {
+    private static final String entityAlreadyExistsTemplate = "Entity: %s already exists";
+
+    public EntityAlreadyExistsException(String entity) {
+        super(String.format(entityAlreadyExistsTemplate, entity));
+    }
+}
