+++ src/main/java/ro/unibuc/hello/repositories/UserRepository.java
@@ -1,0 +1,12 @@
+package ro.unibuc.hello.repositories;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+import ro.unibuc.hello.entities.User;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends MongoRepository<User, String> {
+    Optional<User> findById(String id);
+}
