--- src/test/java/ro/unibuc/hello/dto/GreetingTest.java
@@ -1,19 +1,0 @@
-package ro.unibuc.hello.dto;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-class GreetingTest {
-
-    Greeting myGreeting = new Greeting(1, "John");
-
-    @Test
-    void test_content(){
-        Assertions.assertSame("John", myGreeting.getContent());
-    }
-    @Test
-    void test_id(){
-        Assertions.assertEquals(1, myGreeting.getId());
-    }
-
-}
